Tests

+++ src/main/java/pl/coderstrust/accounting/model/Company.java
@@ -7,7 +7,7 @@ public class Company {
 
   private String companyName;
   private String address;
-  private int taxIdentificationNumber;
+  private long taxIdentificationNumber;
 
   public String getcompanyName() {
     return companyName;
@@ -25,11 +25,11 @@ public class Company {
     this.address = address;
   }
 
-  public int getTaxIdentificationNumber() {
+  public long getTaxIdentificationNumber() {
     return taxIdentificationNumber;
   }
 
-  public void setTaxIdentificationNumber(int taxIdentificationNumber) {
+  public void setTaxIdentificationNumber(long taxIdentificationNumber) {
     this.taxIdentificationNumber = taxIdentificationNumber;
   }
 }
